# refactor GC: do not maintain GCBits of dead objects