There is no need to maintain GCbits for dead objects, so we can avoid clearing them during sweeping (apart from the finalize bits that still need to be cleared because runFinalizers doesn't know which objects are alive).

This PR adds a small overhead in malloc by calling both clrBits and setBits, but this can be reduced by a combined setting function to be added by another PR.