@rainers 2015-01-30
I've run the druntime/benchmarks and they are actually slightly slower after applying this PR.