@andralex 2015-01-30
How much slower? Anyhow your changes make the code smaller which may have an impact on larger codebases.

@MartinNowak LGTY? I'm thinking our two resident GC experts @rainers and @MartinNowak should review each GC-related pull request (and obviously each others').