@rainers 2015-01-30
These are the benchmark on my desktop i7 @ 4GHz
```
              win32 master -> this PR   win64 master -> this PR
R bulk             0.754 s -> 0.813 s        1.981 s -> 2.039 s
R resize           0.758 s -> 0.775 s        0.867 s -> 0.898 s
R string           0.024 s -> 0.024 s        0.022 s -> 0.022 s
R conalloc         0.063 s -> 0.063 s        0.677 s -> 0.668 s
R conappend        0.012 s -> 0.012 s        0.011 s -> 0.011 s
R concpu           0.085 s -> 0.085 s        0.828 s -> 0.822 s
R conmsg           0.621 s -> 0.664 s        0.564 s -> 0.629 s
R dlist            1.404 s -> 1.478 s        1.593 s -> 1.663 s
R huge_single      0.010 s -> 0.010 s        0.008 s -> 0.009 s
R rand_large       3.303 s -> 3.026 s        0.660 s -> 0.677 s
R rand_small       0.623 s -> 0.602 s        0.512 s -> 0.524 s
R slist            1.372 s -> 1.423 s        1.528 s -> 1.597 s
R testgc3          1.023 s -> 1.087 s        1.325 s -> 1.436 s
R tree1            0.701 s -> 0.722 s        0.616 s -> 0.660 s
R tree2            0.805 s -> 0.850 s        0.776 s -> 0.872 s
R words            0.830 s -> 0.817 s        0.796 s -> 0.799 s
```
A few percent slower across most tests. IMO too bad to compensate the code cleanup. I'll try the combined set method...