@rainers 2015-01-30
> I'll try the combined set method...

That doesn't help, so there seems no good rescue for this PR. I guess it's clrBitsSmallSweep that makes the difference.