@MartinNowak 2015-01-31
> That doesn't help, so there seems no good rescue for this PR. I guess it's clrBitsSmallSweep that makes the difference.

Most likely, @dsimcha introduced it, because bitops aren't among the fast operations.
But I actually already have an idea to get rid of clrBitsSmallSweep. Let's me quickly try that.